Rename gem to sidekiq-alive (#8)
