+++ spec/spec_helper.rb
@@ -1,9 +1,10 @@
 require 'bundler/setup'
-require 'sidekiq_alive'
+require 'sidekiq-alive'
 require 'rspec-sidekiq'
 require 'mock_redis'
 require 'rack/test'
 require 'pry'
+
 ENV['RACK_ENV'] = 'test'
 ENV['HOSTNAME'] = 'test-hostname'
 # initialize server
